cleaned up files: alerts was in the wrong branch
