+++ app/index.tsx
@@ -1,6 +1,5 @@
 import { View } from "react-native";
 import { Button, H1, styled } from "tamagui";
-import { ShortAlert } from "../src/components/Alert/ShortAlert";
 
 const CBtn = styled(Button, {
   borderRadius: "$r-strong",
@@ -14,7 +13,6 @@ export default function App() {
     <View>
       <H1>HOME</H1>
       <CBtn>Plain</CBtn>
-      <ShortAlert name='success' shade='subtle' alert="Test" />
     </View>
   );
 }
